requirements.txt
fastapi
uvicorn
twilio
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
apscheduler
python-dotenv
python-multipart